Add check for const classRoute
Check if {{ const classRoute = GCR("view", this) }} exists in router, if not then add it, this gets checked every time a new route is created or updated
